APF-696: Fix AWS cert error handling
Fixes error handling by skipping 404 not found and 400 bad requests when
fetching approval pages.  404 not found is returned when the context param
isn't found.  400 bad request is returned when the approval request is expired.
In both cases, we shouldn't error and instead should filter out the failures
and return the cards that succeeded.

Signed-off-by: John Bard <jbard@vmware.com>