APF-711 Add check for RPM
This change adds a check to ensure that the connectors can be installed
using the generated RPMs.

As part of the build (when the check-rpm profile is active) we now spin
up a Centos 7 Docker container and install and configure the connector.
If the connector's health check doesn't succeed within 60s, the build will fail.